% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argInt2alpha.R
\name{agrInt2alpha}
\alias{agrInt2alpha}
\title{agrInt2alpha}
\usage{
agrInt2alpha(clin.limit, n, sigmae)
}
\arguments{
\item{clin.limit}{Clinically meaningful lower and upper limit}

\item{n}{Sample size}

\item{sigmae}{Variance estimate of residual from measurement error model}
}
\value{
Discordance rate
}
\description{
Function agrInt2alpha calculates discordance rate (alpha) using clinically meaningful limit.
}
\details{
Function agrInt2alpha calculates discordance rate (alpha) using clinically meaningful limit.
}
\examples{
agrInt2alpha(clin.limit=c(-15, 15), n=52, sigmae=46.09245)
}
\references{
Jason J. Z. Liao, Quantifying an Agreement Study, Int. J. Biostat. 2015; 11(1): 125-133
}
\author{
Jialin Xu, Jason Liao
}
