% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPIA}
\alias{IPIA}
\title{IPIA measures from 52 kidneys}
\format{A data frame with 52 rows and 3 variables:
\itemize{
  \item id: sample ids
  \item Urography: IPIA data evaluated by means of computerized urography
  \item Tomography: IPIA data evaluated by means of computerized tomography
}}
\usage{
IPIA
}
\description{
A dataset containing inferior pelvic infundibular angle (IPIA) dataset measured by urography and tomography on n=52 kidneys. The variables are as follows:
}
\references{
Luiz RR, Costa AJL, Kale PL, Werneck GL. Assessment of agreement of a quantitative variable: a new graphical approach. J Clin Epidemiol 2003; 56:963-7.
}
\keyword{datasets}
