% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ai.R
\name{summary.ai}
\alias{summary.ai}
\title{summary.ai}
\usage{
\method{summary}{ai}(object, ...)
}
\arguments{
\item{object}{ai object from ai function}

\item{...}{additional arguments affecting the summary produced}
}
\value{
Function summary.ai prints out key summaries on screen
}
\description{
The summary method for ai objects
}
\examples{
a <- c(1, 2, 3, 4, 7)
b <- c(1, 3, 2, 5, 3)
ans <- ai(x=a, y=b)
summary(ans)
}
\references{
Jason J. Z. Liao, Quantifying an Agreement Study, Int. J. Biostat. 2015; 11(1): 125-133
}
\author{
Jialin Xu, Jason Liao
}
